Local Project Development - Version 1
Mainly Created this as a passion project and needed to practice some python. The goal for this project is to create a Dndbeyond type character sheet. This first version is a MVP to showcase that its possible. I will probably try the next version in python and then switch to Java + Javascript + HTML + CSS in a few version at least once I get the local version a bit better.